Refactored to reduce data logic and renamed
