+++ src/components/sightingcard/index.js
@@ -1,0 +1,3 @@
+import SightingCardList from './SightingCardList';
+
+export default SightingCardList;
