+++ src/components/sightingcard/style.scss
