--- src/components/sightingtable/index.js
@@ -1,3 +1,0 @@
-import SightingTable from './SightingTable';
-
-export default SightingTable;
